+++ resources/js/components/modals/attribute/Add.vue
@@ -58,9 +58,13 @@
                     </h5>
                     <attribute-list
                         :attributes="state.previewAttribute.attribute"
+                        :disable-drag="true"
                         :options="{'hide_labels': true}"
+                        :nolabels="true"
                         :values="state.previewAttribute.values"
                         :selections="{}"
+                        :preview="true"
+                        :preview-data="state.previewData"
                     />
                 </div>
             </div>
@@ -92,6 +96,10 @@
         reactive,
     } from 'vue';
     import { useI18n } from 'vue-i18n';
+
+    import {
+        randomId,
+    } from '@/helpers/helpers.js';
 
     import AttributeTemplate from '@/components/AttributeTemplate.vue';
 
@@ -187,6 +195,7 @@
             };
 
             // DATA
+            const fakeId = randomId();
             const state = reactive({
                 attribute: {},
                 columns: [],
@@ -200,8 +209,6 @@
                 previewAttribute: computed(_ => {
                     if(!state.attribute.type) return null;
 
-                    const fakeId = Number.MAX_VALUE;
-
                     return {
                         attribute: [{
                             ...state.attribute,
@@ -214,6 +221,30 @@
                                 value: getInitialValue(state.attribute),
                             },
                         },
+                    };
+                }),
+                previewData: computed(_ => {
+                    if(!state.attribute.type) return null;
+                    // Currently only table type supports additional data
+                    if(state.attribute.type != 'table') return null;
+
+                    const columns = {};
+                    if(state.columns.length == 0) {
+                        columns.is_preview = true;
+                    } else {
+                        state.columns.forEach((c, i) => {
+                            const colId = fakeId + (i+1);
+                            columns[colId] = {
+                                id: colId,
+                                parent_id: fakeId,
+                                recursive: c.recursive,
+                                datatype: c.type,
+                                thesaurus_url: c.label.concept_url
+                            };
+                        });
+                    }
+                    return {
+                        [fakeId]: columns,
                     };
                 }),
             });